handle user not login vote dan add getters store

+++ client/src/components/VoteAnswer.vue
@@ -3,6 +3,21 @@
     <a class="like" @click="voteAnswer">
       <i class="thumbs outline up icon"></i> {{votes.length}} Likes
     </a>
+    <div :class="errorClass">
+      <div class="ui icon header" style="color: red;">
+        <i class="remove circle outline icon"></i>
+        Access Denied!!
+      </div>
+      <div class="content">
+        <p style="color: red; text-align: center; font-size: 18px;">Please login before response any questions or answer!!</p>
+      </div>
+      <div class="actions">
+        <div class="ui red basic cancel inverted button">
+          <i class="remove icon"></i>
+          Close
+        </div>
+      </div>
+    </div>
   </div>
 </template>
 
@@ -11,6 +26,7 @@ export default {
   props: ['answerId'],
   data () {
     return {
+      errorClass: `ui basic modal ${this.questionId}`,
       votes: []
     }
   },
@@ -18,20 +34,32 @@ export default {
     voteAnswer () {
       let index = this.votes.map(v => { return v.userId }).indexOf(localStorage.getItem('userId'))
       // console.log(index)
-      // if (index == -1) {
-      this.$http.post('/voteAnswers', {
-        answerId: this.answerId,
-        votes: true
-      }, {
-        headers: {
-          accesstoken: localStorage.getItem('accesstoken')
-        }
-      })
-      .then(({data}) => {
-        this.votes.push(data)
-      })
-      // eslint-disable-next-line
-      .catch(err => {
+      if (index == -1) {
+        this.$http.post('/voteAnswers', {
+          answerId: this.answerId,
+          votes: true
+        }, {
+          headers: {
+            accesstoken: localStorage.getItem('accesstoken')
+          }
+        })
+        .then(({data}) => {
+          if(data.access) {
+            $(`.ui.basic.modal.${this.questionId}`)
+              .modal('show')
+            ;
+            console.log('error')
+          } else {
+            this.votes.push(data)
+          }
+        })
+        // eslint-disable-next-line
+        .catch(err => {
+          $(`.ui.basic.modal.${this.questionId}`)
+            .modal('show')
+          ;
+        })
+      } else {
         this.$http.delete(`/voteAnswers/${this.votes[index]._id}`, {
           headers: {
             accesstoken: localStorage.getItem('accesstoken')
@@ -41,9 +69,8 @@ export default {
           this.votes.splice(index, 1)
         })
         .catch(error => console.log(error))
-      })
-      // } else {
       // console.log(this.votes[index]._id)
+      }
     },
     getVotes () {
       this.$http.get(`/voteAnswers/answers/${this.answerId}`)
