+++ client/src/components/VoteQuestion.vue
@@ -3,6 +3,25 @@
     <a class="like" @click="voteAnswer">
       <i class="thumbs outline up icon"></i> {{votes.length}} Likes
     </a>
+    <div :class="errorClass">
+      <div class="ui icon header" style="color: red;">
+        <i class="remove circle outline icon"></i>
+        Access Denied!!
+      </div>
+      <div class="content">
+        <p style="color: red; text-align: center; font-size: 18px;">Please login before response any questions or answer!!</p>
+      </div>
+      <div class="actions">
+        <div class="ui red basic cancel inverted button">
+          <i class="remove icon"></i>
+          Close
+        </div>
+        <router-link to="/login" class="ui green ok inverted button">
+          <i class="checkmark icon"></i>
+          Login
+        </router-link>
+      </div>
+    </div>
   </div>
 </template>
 
@@ -12,27 +31,49 @@ export default {
   props: ['questionId'],
   data () {
     return {
-      votes: []
+      errorClass: `ui basic modal ${this.questionId}`,
+      votes: [],
+      isLogin: localStorage.getItem('accesstoken')
     }
   },
   methods: {
     voteAnswer () {
       let index = this.votes.map(v => { return v.userId }).indexOf(localStorage.getItem('userId'))
-      // console.log(index)
-      // if (index == -1) {
-      this.$http.post('/voteQuestions', {
-        questionId: this.questionId,
-        votes: true
-      }, {
-        headers: {
-          accesstoken: localStorage.getItem('accesstoken')
-        }
-      })
-      .then(({data}) => {
-        this.votes.push(data)
-      })
-      .catch(err => {
-        // console.log(err)
+      console.log(index)
+      if (index == -1) {
+        this.$http.post('/voteQuestions', {
+          questionId: this.questionId,
+          votes: true
+        }, {
+          headers: {
+            accesstoken: localStorage.getItem('accesstoken')
+          }
+        })
+        .then(({data}) => {
+          if(data.access) {
+            $(`.ui.basic.modal.${this.questionId}`)
+              .modal('show')
+            ;
+            console.log('error')
+          } else {
+            this.votes.push(data)
+          }
+        })
+        .catch(err => {
+          $(`.ui.basic.modal.${this.questionId}`)
+            .modal('show')
+          ;
+          // this.$http.delete(`/voteQuestions/${this.votes[index]._id}`, {
+          //   headers: {
+          //     accesstoken: localStorage.getItem('accesstoken')
+          //   }
+          // })
+          // .then(response => {
+          //   this.votes.splice(index, 1)
+          // })
+          // .catch(error => console.log(error))
+        })
+      } else {
         this.$http.delete(`/voteQuestions/${this.votes[index]._id}`, {
           headers: {
             accesstoken: localStorage.getItem('accesstoken')
@@ -42,9 +83,7 @@ export default {
           this.votes.splice(index, 1)
         })
         .catch(error => console.log(error))
-      })
-      // } else {
-      // console.log(this.votes[index]._id)
+      }
     },
     getVotes (questionId) {
       this.$http.get(`/voteQuestions/questions/${questionId}`)
